hack: linking flow demo complete
